Add codes and test data and report
The report is in the Doc folder, the source_code have three parts: one is SavedModels which contains the models already trained and can be used in evaluation part derectly. It will save your time. The other is the shiny app codes including glob, server, ui three parts. The last one is .csv file which is the dataset used to find the best regression model. The shiny code shows the interaction more intuitively, that is the recommand one. User can do some exploratory with the interface. That is awesome!